Move track logic into a module, fix isolate() first split
